Fix update offers & clean up
