+++ src/model/points.js
@@ -1,5 +1,4 @@
 import AbstractObserver from '../utils/abstract-observer.js';
-import {generateRandomString} from '../utils/common';
 
 export default class Points extends AbstractObserver {
   constructor() {
@@ -72,11 +71,6 @@ export default class Points extends AbstractObserver {
           description: point.destination['description'],
           pictures: point.destination['pictures'],
         },
-        offers: point.offers.map((offer) => ({
-          id: generateRandomString(5, 9),
-          title: offer['title'],
-          price: offer['price'],
-        })),
       },
     );
 
@@ -86,7 +80,6 @@ export default class Points extends AbstractObserver {
     delete adaptedTask['date_to'];
     delete adaptedTask['is_favorite'];
     delete adaptedTask.destination['name'];
-    delete adaptedTask.offers.map((offer) => offer.id);
 
     return adaptedTask;
   }
@@ -105,10 +98,6 @@ export default class Points extends AbstractObserver {
           'description': point.destination.description,
           'pictures': point.destination.pictures,
         },
-        offers: point.offers.map((offer) => ({
-          title: offer['title'],
-          price: offer['price'],
-        })),
       },
     );
