+++ src/presenter/point-new.js
@@ -1,7 +1,6 @@
 import PointItemEditView from '../views/point-item-modify';
 import {remove, render, RenderPosition} from '../utils/render.js';
 import {UserAction, UpdateType} from '../const.js';
-import {generateRandomString} from '../utils/common';
 import {ITEM_BLANK} from '../const';
 
 export default class PointNew {
@@ -47,9 +46,7 @@ export default class PointNew {
     this._changeData(
       UserAction.ADD_POINT,
       UpdateType.MAJOR,
-      // Пока у нас нет сервера, который бы после сохранения
-      // выдывал честный id задачи, нам нужно позаботиться об этом самим
-      Object.assign({id: generateRandomString(8, 15)}, point),
+      point,
     );
     this.destroy();
   }
